move the forcing scheme funcs to lbm.py and fix a serious error in source term compuation
